Improve pipelined file transfer logic
Before this change, if the reqack or ack was clobbered, the sender would
re-send the entire round, plus the reqack.  Thus, if just the reqack,
ack (or maybe the last block in the round plus the req) was clobbered,
we would re-send the entire round unnecessarily.  This also means that
if the remote goes away, we send the entire round ten times until we
decide they're gone, which uses a bunch of airtime, power, and heat for
no reason.

This change improves the logic to just resend the reqack.  Thus, if
they got none of it, they'll respond with nothing.  If they got some or
all, they'll respond and we'll only resend what we need to.  This should
improve performance a bit too, because we resend less data and recover
more quickly.