Massively simplify the handling of GPS/GPS-A icons.
In preferences, there is now a single graphical default icon selection tool. This icon is used
for GPS and GPS-A QSTs, as well as remotely-requested position beacons (which have never included
an icon before). The default comment is used for GPS and GPS-A beacons when requested, and provides
the default input for new QSTs of both types.